+++ python_setuptools/{{cookiecutter.project_slug}}/CHANGELOG.md
@@ -6,3 +6,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ## [{{ cookiecutter.version}}] - ({% now 'local', '%Y-%m-%d' %})
+- First version created by a cookiecutter.